refactor: introduce test runner
